Refactor chat UI directory structure
