+++ apps/web/src/app/layout.tsx
@@ -17,8 +17,7 @@ import { getUserSettings } from "@/server/userSettings";
 import { extractUserIdFromHeaders, extractWorkspaceIdFromHeaders } from "@/server/userId";
 import { resolveWorkspaceForCurrentRequestIdentity } from "@/server/workspaceBootstrap";
 import { AccountMenu } from "@/ui/AccountMenu";
-import { ChatLayoutProvider } from "@/ui/chat/ChatLayoutProvider";
-import { ChatLayoutShell } from "@/ui/chat/ChatLayoutShell";
+import { ChatLayoutProvider, ChatLayoutShell } from "@/ui/chat";
 import { CurrencySelector } from "@/ui/CurrencySelector";
 import { FilteredBanner } from "@/ui/FilteredBanner";
 import { FilteredModeProvider } from "@/ui/FilteredModeProvider";
