+++ apps/web/src/ui/chat/components/ChatLayoutShell.tsx
