+++ apps/web/src/ui/chat/components/ChatPanelHeader.tsx
