+++ apps/web/src/ui/chat/components/FileAttachment.tsx
