+++ apps/web/src/ui/chat/display/chatTranscriptMarkdown.ts
