+++ apps/web/src/ui/chat/display/thinkingSummary.ts
