+++ apps/web/src/ui/chat/display/toolCallDisplay.ts
