+++ apps/web/src/ui/chat/hooks/chatAutoScroll.ts
