+++ apps/web/src/ui/chat/hooks/useDesktopEnterToSend.ts
