+++ apps/web/src/ui/chat/index.ts
@@ -1,0 +1,5 @@
+"use client";
+
+export { ChatLayoutProvider } from "./components/ChatLayoutProvider";
+export { ChatLayoutShell } from "./components/ChatLayoutShell";
+export { ChatPanel } from "./components/ChatPanel";
