+++ apps/web/src/ui/chat/session/chatSessionSnapshot.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import type { StoredMessage } from "@/ui/hooks/useChatHistory";
-import type { ChatRunState } from "./streamRecovery";
+import type { ChatRunState } from "../stream/streamRecovery";
 
 export type ChatSessionSnapshot = Readonly<{
   sessionId: string;
